Moved csrf-token related functions to utils for better project structure. Removed some unused imports and refactored characte_routes to include a missing error return.
