+++ src/services/user_service.py
@@ -4,6 +4,7 @@ from sqlalchemy import text
 from sqlalchemy.exc import IntegrityError
 from werkzeug.security import check_password_hash, generate_password_hash
 
+from utils.csrf_utils import get_csrf_token
 from db.db import db
 
 
@@ -76,18 +77,3 @@ def logout_user():
             del session[key]
         except KeyError:
             pass
-
-
-def get_csrf_token():
-    if "csrf_token" not in session:
-        create_csrf_token()
-    return session["csrf_token"]
-
-
-def create_csrf_token():
-    session["csrf_token"] = os.urandom(16).hex()
-
-
-def check_csrf():
-    if session["csrf_token"] != request.form["csrf_token"]:
-        abort(403)
